requirements.txt
Werkzeug==0.9.4
args==0.1.0
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.3.2
bitly-api==0.3
bleach==1.4
clint==0.3.4
disqus-python==0.4.2
feedparser==5.1.3
geopy==0.97
html5lib==1.0b3
httplib2==0.8
logging==0.4.9.6
mongoengine==0.8.7
oauth2==1.5.211
oauthlib==0.6.1
pymongo==2.6.3
python-postmark==0.3.2
python-twitter==1.3.1
requests==2.2.1
requests-oauthlib==0.4.0
selenium==2.40.0
simplejson==3.3.3
six==1.5.2
slugify==0.0.1
splinter==0.6.0
sunlight==1.2.1
tornado==3.2
tweepy==2.2
wsgiref==0.1.2
